--- admin/services/tech-support/ps-report-exec.service/ps-report-exec.service.ts
@@ -1,6 +1,0 @@
-
-export class PsReportExecService {
-  public static async requestReportGeneration (currentUserId: number): Promise<any> {
-    throw new Error('currentUserId must be a non negative number')
-  }
-}
